first pass at removing equations from the model diagram
